Revert "Adds extinguishers to various atmos suit storage slots and fixes fire…"
This reverts commit d6e05aa40b6014bfd813633ccec95659d950ae1b.

+++ code/game/objects/effects/effect_system/effects_foam.dm
@@ -42,7 +42,7 @@
 
 	var/turf/open/T = get_turf(src)
 	var/obj/effect/hotspot/hotspot = (locate(/obj/effect/hotspot) in T)
-	if(hotspot && T.air)
+	if(hotspot && istype(T) && T.air)
 		qdel(hotspot)
 		var/datum/gas_mixture/G = T.air
 		var/plas_amt = min(30,GET_MOLES(/datum/gas/plasma, G)) //Absorb some plasma
